Separated the code in script.js to more files
