+++ frontend/index.html
@@ -55,6 +55,6 @@
         </div>
       </div>
     </main>
-    <script type="module" src="./script.js"></script>
+    <script type="module" src="./js/app.js"></script>
   </body>
 </html>
